pyproject.toml
[tool.poetry]
name = "workflows_with_litellm"
version = "0.1.1"
description = ""
authors = ["Shuyi Wang <nkwshuyi@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pyperclip = "^1.9.0"
litellm = "^1.51.0"
langchain = "^0.2.7"
python-dotenv = "^1.0.1"
openai = "^1.35.14"
generalagent = "^0.3.23"
exa-py = "^1.0.16"
schema = "^0.7.7"
streamlit = "^1.38.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
